requirements.txt
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
numpy>=1.26.0
scipy>=1.11.0
tqdm>=4.66.0
pillow>=10.0.0
opencv-python>=4.8.0.76
pyyaml>=6.0.1
einops>=0.8.0
lpips>=0.1.4
ptflops>=0.7.3
thop>=0.1.1-post2209072238
tensorboard>=2.17.0
h5py>=3.11.0
pre-commit>=3.7.0
pytest>=8.0.0
